refactor(injection): 优化管线，将轮次记忆去重逻辑转移至ABMCollector
